ToDoList/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
future==1.0.0
serializers==0.2.4
sqlparse==0.5.1
